ajax login

+++ service/src/main/java/com/tipcrm/service/UserService.java
@@ -3,7 +3,7 @@ import java.util.List;
 import java.util.Set;
 
 public interface UserService {
-    Set<String> getRoleListByUserId(Long userId);
+    Set<String> getRoleListByUserId(Integer userId);
 
-    Set<String> getPermissionValueListByUserId(Long userId);
+    Set<String> getPermissionValueListByUserId(Integer userId);
 }
